move lookup data preparation to amberdb instantiation.
